Change abstract radiative properties to be its own module
